refactor: remove circular dependency
