+++ src/elements/Circle.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { extract, stringifyPropsForFabric } from '../lib/extract/extractProps';
 import { CommonPathProps, NumberProp } from '../lib/extract/types';
 import Shape from './Shape';
-import { RNSVGCircle } from '../ReactNativeSVG';
+import RNSVGCircle from '../fabric/CircleNativeComponent';
 import { NativeMethods } from 'react-native';
 
 export interface CircleProps extends CommonPathProps {
