+++ src/elements/Defs.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { RNSVGDefs } from '../ReactNativeSVG';
+import RNSVGDefs from '../fabric/DefsNativeComponent';
 
 export default class Defs extends Component<React.PropsWithChildren<{}>> {
   static displayName = 'Defs';
