+++ src/elements/RadialGradient.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from 'react';
 import extractGradient from '../lib/extract/extractGradient';
 import { NumberProp, TransformProps, Units } from '../lib/extract/types';
 import Shape from './Shape';
-import { RNSVGRadialGradient } from '../ReactNativeSVG';
+import RNSVGRadialGradient from '../fabric/RadialGradientNativeComponent';
 import { stringifyPropsForFabric } from '../lib/extract/extractProps';
 import { NativeMethods } from 'react-native';
