+++ src/elements/Text.tsx
@@ -12,7 +12,7 @@ import {
 import { pickNotNil } from '../lib/util';
 import Shape from './Shape';
 import './TSpan';
-import { RNSVGText } from '../ReactNativeSVG';
+import RNSVGText from '../fabric/TextNativeComponent';
 
 export interface TextProps extends TextSpecificProps {
   children?: ReactNode;
